packages/calypso-e2e/src/rest-api-client.ts - implement new method to see if there are posts on the site.
test/e2e/specs/tools/advertising__promote.ts
- stop creating new post unnecessarily

+++ test/e2e/specs/tools/advertising__promote.ts
@@ -50,15 +50,22 @@ skipDescribeIf( envVariables.ATOMIC_VARIATION === 'private' )(
 			const accountName = getTestAccountByFeature( envToFeatureKey( envVariables ) );
 			testAccount = new TestAccount( accountName );
 
-			// Createa a new test post before starting the test, to ensure at least one
-			// available post.
 			restAPIClient = new RestAPIClient( testAccount.credentials );
-			newPostDetails = await restAPIClient.createPost(
+
+			// Createa a new test post before starting the test if site has no published post.
+			const hasPosts = await restAPIClient.siteHasPost(
 				testAccount.credentials.testSites?.primary.id as number,
-				{
-					title: pageTitle,
-				}
+				{ state: 'publish' }
 			);
+
+			if ( ! hasPosts ) {
+				newPostDetails = await restAPIClient.createPost(
+					testAccount.credentials.testSites?.primary.id as number,
+					{
+						title: pageTitle,
+					}
+				);
+			}
 
 			await testAccount.authenticate( page );
 
@@ -75,11 +82,12 @@ skipDescribeIf( envVariables.ATOMIC_VARIATION === 'private' )(
 		} );
 
 		it( 'Click on Promote for the first post', async function () {
-			await advertisingPage.clickButtonByNameOnRow( 'Promote', { postTitle: pageTitle } );
+			await advertisingPage.clickButtonByNameOnRow( 'Promote', { row: 1 } );
 		} );
 
 		it( 'Land in Blaze campaign landing page', async function () {
 			await page.waitForURL( /advertising/ );
+
 			blazeCampaignPage = new BlazeCampaignPage( page );
 		} );
 
@@ -102,6 +110,10 @@ skipDescribeIf( envVariables.ATOMIC_VARIATION === 'private' )(
 		} );
 
 		afterAll( async function () {
+			if ( ! newPostDetails ) {
+				return;
+			}
+
 			await restAPIClient.deletePost(
 				testAccount.credentials.testSites?.primary.id as number,
 				newPostDetails.ID
